requirements.txt
Flask==3.0.0
biopython==1.81
pandas==2.1.1
scikit-learn==1.3.1
